projects/experiment-tracker/experiment_tracker/parser/markdown_parser.py
"""Markdown experiment file parser."""

import re
from typing import Dict, Any, List, Optional
from pathlib import Path
import yaml
import markdown


class MarkdownExperimentParser:
    """
    Parse ADHD experiment markdown files.

    Extracts frontmatter metadata and structured metrics
    from markdown files following the experiment template format.
    """

    def __init__(self):
        """Initialize parser."""
        self.md = markdown.Markdown(extensions=['meta'])

    def parse_file(self, file_path: str) -> Dict[str, Any]:
        """
        Parse a single experiment markdown file.

        Args:
            file_path: Path to markdown file

        Returns:
            Dictionary with parsed experiment data:
            - metadata: Frontmatter data (title, date, type, status)
            - metrics: List of daily metrics
            - observations: Extracted observations
            - learnings: Extracted learnings
            - next_steps: Extracted next steps
        """
        # TODO: Implement file parsing
        # 1. Read file content
        # 2. Extract YAML frontmatter
        # 3. Parse markdown body
        # 4. Extract metrics from "Metrics" section
        # 5. Extract observations, learnings, next steps
        pass

    def extract_frontmatter(self, content: str) -> Dict[str, Any]:
        """
        Extract YAML frontmatter from markdown content.

        Args:
            content: Raw markdown content

        Returns:
            Dictionary of frontmatter data
        """
        # TODO: Implement frontmatter extraction
        pass

    def extract_metrics(self, content: str) -> List[Dict[str, Any]]:
        """
        Extract daily metrics from markdown content.

        Looks for patterns like:
        - **Date**: 2024-01-15
        - **Time-Boxed Sessions**: 6
        - **Energy Level**: 4/5
        etc.

        Args:
            content: Markdown content

        Returns:
            List of metric dictionaries, one per day
        """
        # TODO: Implement metric extraction
        # Use regex to find metric patterns
        # Parse different formats (numbers, ratios, percentages)
        # Return structured data
        pass

    def parse_metric_value(self, raw_value: str, metric_name: str) -> Any:
        """
        Parse a metric value from string to appropriate type.

        Handles:
        - Numbers: "6" -> 6
        - Ratios: "4/5" -> 0.8
        - Percentages: "80%" -> 0.8
        - Ratings: "4/5" -> 4.0 (for energy/joy)

        Args:
            raw_value: Raw string value
            metric_name: Name of the metric (for context)

        Returns:
            Parsed value in appropriate type
        """
        # TODO: Implement value parsing
        pass

    def validate_metrics(self, metrics: Dict[str, Any]) -> bool:
        """
        Validate that required metrics are present.

        Args:
            metrics: Metrics dictionary

        Returns:
            True if valid, False otherwise
        """
        required_fields = [
            "date",
            "Time-Boxed Sessions",
            "Tasks Completed",
            "Energy Level",
            "Joy Rating"
        ]

        # TODO: Implement validation
        pass
