projects/experiment-tracker/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-markdown>=3.5.0
pyyaml>=6.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Visualization
plotly>=5.17.0
rich>=13.0.0
jinja2>=3.1.0
kaleido>=0.2.1  # For static image export

# Analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# AI Insights (Optional)
openai>=1.0.0
anthropic>=0.8.0

# CLI
typer>=0.9.0
click>=8.1.0

# Utilities
httpx>=0.25.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
